change file names to delta_decode and delta_encode
