+++ chainercv/links/__init__.py
@@ -1,7 +1,7 @@
 from chainercv.links.model.faster_rcnn.faster_rcnn import FasterRCNNBase  # NOQA
 from chainercv.links.model.faster_rcnn.faster_rcnn_vgg import FasterRCNNVGG16  # NOQA
 from chainercv.links.model.faster_rcnn.region_proposal_network import RegionProposalNetwork  # NOQA
-from chainercv.links.model.faster_rcnn.utils.bbox_regression_target import bbox_regression_target  # NOQA
-from chainercv.links.model.faster_rcnn.utils.bbox_regression_target import bbox_regression_target_inv  # NOQA
+from chainercv.links.model.faster_rcnn.utils.delta_decode import delta_decode  # NOQA
+from chainercv.links.model.faster_rcnn.utils.delta_encode import delta_encode  # NOQA
 from chainercv.links.model.faster_rcnn.utils.generate_anchor_base import generate_anchor_base  # NOQA
 from chainercv.links.model.faster_rcnn.utils.proposal_creator import ProposalCreator  # NOQA
